+++ tests/foreign_keys/m2m_string/test_many_to_many_no_related_name.py
@@ -14,6 +14,7 @@ models = edgy.Registry(database=database)
 
 class Album(edgy.Model):
     id = edgy.IntegerField(primary_key=True, autoincrement=True)
+    name = edgy.fields.CharField(max_length=100)
     tracks = edgy.ManyToMany("Track", related_name=False)
 
     class Meta: