+++ tests/models/run_sync/test_model_primary_key.py
@@ -56,7 +56,7 @@ async def test_model_default_primary_key():
 
 
 async def test_model_custom_primary_key():
-    profile = edgy.run_sync(Profile.query.create(name="Test", language="EN", age=18))
+    profile = edgy.run_sync(Profile.query.create(language="EN", age=18))
     profiles = edgy.run_sync(Profile.query.filter())
 
     assert len(profiles) == 1